Adicionado validação para funcionario não poder se remover

+++ Services/Services/Implementation/DepartamentoService.cs
@@ -117,9 +117,14 @@ namespace Services.Services.Implementation
             return _mapper.Map<List<DepartamentoDTO>>(departamentos);
         }
 
-        public void RemoveFuncionarioDepartamento(Guid departamentoId, List<Guid> funcionarioId)
+        public void RemoveFuncionarioDepartamento(Guid departamentoId, List<Guid> funcionarioId, Guid funcionarioLogadoId)
         {
             var funcionarioDepartamento = _repository.DepartamentoFuncionarioRepository.FindByCondition(x => x.DepartamentoId == departamentoId && funcionarioId.Any(y => y == x.FuncionarioId)).ToList();
+
+            if(funcionarioDepartamento.Any(x => x.FuncionarioId == funcionarioLogadoId))
+            {
+                throw new FuncionarioNaoPodeSeRemover();
+            }
 
             _repository.DepartamentoFuncionarioRepository.DeleteMultiple(funcionarioDepartamento);
             _repository.Save();
